requirements.txt
langchain>=0.0.267
langchain-community>=0.0.13
langchain-ollama>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.15.1
python-dotenv>=1.0.0
tqdm>=4.66.1
tabulate>=0.9.0
requests>=2.28.0
